Add period health data API + glucose type change
Add a GET endpoint to fetch all health data for a user within a date range and corresponding service method GetHealthDataForPeriodAsync that queries each concrete type separately for performance. Introduce HealthDataResponse DTO and extend MapperFactory, NumericMapper, WorkoutMapper and ManualInputMapper with ToDto mappings and helper maps to convert entities to DTOs. Change GlucoseLevel.mmolL from double to decimal and set precision in DiabitsDbContext; update model snapshot and add two EF Core migrations for precision and type changes. Misc: initialize HealthValue properties in NumericDto/WorkoutDto, adjust end-of-day logic in GetManualInputForDayAsync, and reorder/restore RefreshToken index in DbContext.

+++ Diabits.API/Models/HealthDataPoints/HealthConnect/GlucoseLevel.cs
@@ -5,5 +5,5 @@
 /// </summary>
 public class GlucoseLevel : HealthDataPoint
 {
-    public double mmolL { get; set; }
+    public decimal mmolL { get; set; }
 }
